[refactor] background 리퀘스트에 따라 파일 분할
